requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
boto3==1.34.0
botocore==1.34.0
requests==2.31.0
pypdf2==3.0.1
pdfplumber==0.10.3
python-docx==1.1.0
pillow==10.1.0
pytesseract==0.3.10
spacy==3.7.2
nltk==3.8.1
redis==5.0.1
celery==5.3.4
python-dotenv==1.0.0
jinja2==3.1.2
aiofiles==23.2.1